# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Fluxa API"
  version: "1.0.0"
  description: |
    Fluxa is a multi-currency digital asset platform built on Stellar. It provides
    wallets, instant transfers, FX conversions, scheduled recurring payments,
    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.

    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.
  contact:
    name: "Fluxa Engineering"
    email: "engineering@fluxa.io"
    url: "https://github.com/fluxa/fluxa"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: http://localhost:3000
    description: "Local development"
  - url: https://api.fluxa.io
    description: "Production"
...
security:
  - BearerAuth: []
paths:
...
  /v1/transfers/batch:
    post:
      operationId: createBatch
      summary: "Create a batch transfer"
      description: |
        Submits up to 100 transfers from a single source wallet in one request.
        Each transfer is processed independently; partial success is possible.
      tags: ["Transfers"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchRequest"
            example:
              from_wallet_id: "w_1a2b3c4d"
              transfers:
                - to_wallet_id: "w_5e6f7g8h"
                  asset: "USDC"
                  amount: "50.0000000"
                  reference: "payroll-jan"
                - to_wallet_id: "w_9i0j1k2l"
                  asset: "USDC"
                  amount: "75.0000000"
                  reference: "payroll-jan"
      responses:
        "202":
          description: "Batch accepted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/transfers/batch/{batchId}:
    get:
      operationId: getBatch
      summary: "Get batch status"
      description: "Returns the batch and the status of every individual transfer in it."
      tags: ["Transfers"]
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
          description: "Batch ID"
      responses:
        "200":
          description: "Batch details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/transfers/batch/{batchId}/export:
    get:
      operationId: exportBatch
      summary: "Export batch as CSV"
      description: "Downloads the batch transfer results as a CSV file."
      tags: ["Transfers"]
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
          description: "Batch ID"
      responses:
        "200":
          description: "CSV file"
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── FX ────────────────────────────────────────────────────────────────────
...
  /v1/admin/compliance/reviews:
    get:
      operationId: listComplianceReviews
      summary: "List compliance reviews"
      description: |
        Lists transfers held by compliance screening. A held transfer is
        persisted with status `compliance_hold` and is deliberately not
        enqueued for settlement until it is approved.
      tags: ["Admin"]
      security:
        - AdminAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: ["pending", "approved", "rejected"]
          description: "Filter by review status"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: "Compliance reviews"
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: "#/components/schemas/ComplianceReview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/compliance/reviews/{id}:
    get:
      operationId: getComplianceReview
      summary: "Get a compliance review"
      tags: ["Admin"]
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Compliance review"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "Compliance review not found"
  /v1/admin/compliance/reviews/{id}/approve:
    post:
      operationId: approveComplianceReview
      summary: "Approve a held transfer"
      description: |
        Releases a held transfer. The transaction is reset to `pending` and
        then re-enqueued for settlement — in that order, because the
        settlement engine skips any transaction that is not `pending`.

        Decisions are terminal: approving an already-decided review returns
        `409 REVIEW_ALREADY_DECIDED`, which is what stops two reviewers from
        releasing the same payment twice.
      tags: ["Admin"]
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  description: "Reviewer notes recorded on the audit trail"
      responses:
        "200":
          description: "Review approved and transfer released"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "Compliance review not found"
        "409":
          description: "Review has already been decided"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "REVIEW_ALREADY_DECIDED"
                  message: "compliance review has already been decided"
  /v1/admin/compliance/reviews/{id}/reject:
    post:
      operationId: rejectComplianceReview
      summary: "Reject a held transfer"
      description: |
        Rejects a held transfer. The transaction is marked `failed` and is
        never enqueued for settlement.
      tags: ["Admin"]
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
      responses:
        "200":
          description: "Review rejected"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "Compliance review not found"
        "409":
          description: "Review has already been decided"
  /v1/admin/compliance/sanctions-status:
    get:
      operationId: getSanctionsStatus
      summary: "Get sanctions list status"
      description: |
        Reports the state of the in-memory OFAC SDN set and the outcome of the
        last refresh. `loaded: false` means screening is failing closed and
        every transfer is being held for review.
      tags: ["Admin"]
      security:
        - AdminAuth: []
      responses:
        "200":
          description: "Sanctions list status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SanctionsStatus"
              example:
                loaded: true
                entity_count: 17423
                address_count: 621
                name_count: 16802
                updated_at: "2026-01-15T03:00:12Z"
                last_refresh:
                  status: "success"
                  entity_count: 17423
                  duration_ms: 8412
                  finished_at: "2026-01-15T03:00:12Z"

# ══════════════════════════════════════════════════════════════════════════════
# Components
# ══════════════════════════════════════════════════════════════════════════════
components:
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass a tenant API key as `Bearer <key>` in the `Authorization` header.
        Keys are prefixed with `sk_live_` (production) or `sk_test_` (sandbox).
    AdminAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass an admin API key as `Bearer <key>`. Admin keys are prefixed with
        `sk_admin_`.

  # ── Parameters ──────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "BAD_REQUEST"
              message: "invalid request body"
    Unauthorized:
      description: "Unauthorized — missing or invalid API key"
      content:
        application/json:
          schema:
            type: object
          example:
            error: "missing or invalid authorization header"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "NOT_FOUND"
              message: "resource not found"
    Forbidden:
      description: "Forbidden — admin access required"
      content:
        application/json:
          schema:
            type: object
          example:
            error:
              code: "FORBIDDEN"
              message: "admin access required"
...
  schemas:
    ComplianceReview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transaction_id:
          type: string
          format: uuid
        status:
          type: string
          enum: ["pending", "approved", "rejected"]
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
        rules_fired:
          type: array
          items:
            type: string
            enum:
              - "sanctions_address_match"
              - "sanctions_name_match"
              - "sanctions_name_fuzzy_match"
              - "velocity_burst"
              - "structuring"
              - "round_trip"
              - "screener_error"
        reason:
          type: string
        reviewed_by:
          type: string
          description: "Empty for approvals made with API-key authentication, which carries no user identity."
        review_notes:
          type: string
        reviewed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    SanctionsStatus:
      type: object
      properties:
        loaded:
          type: boolean
          description: "False means screening is failing closed and holding every transfer."
        entity_count:
          type: integer
        address_count:
          type: integer
        name_count:
          type: integer
        updated_at:
          type: string
          format: date-time
        last_refresh:
          type: object
          properties:
            status:
              type: string
              enum: ["success", "failed"]
            entity_count:
              type: integer
            duration_ms:
              type: integer
            error:
              type: string
            finished_at:
              type: string
              format: date-time

    # ── Error ────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: ["error"]
      properties:
        error:
          type: object
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - "BAD_REQUEST"
                - "NOT_FOUND"
                - "UNAUTHORIZED"
                - "FORBIDDEN"
                - "INTERNAL_ERROR"
                - "QUOTE_EXPIRED"
                - "QUOTE_ALREADY_USED"
              description: "Machine-readable error code"
            message:
              type: string
              description: "Human-readable error message"

    # ── Health ────────────────────────────────────────────────────────────
...
    CreateBatchRequest:
      type: object
      required: ["from_wallet_id", "transfers"]
      properties:
        from_wallet_id:
          type: string
          format: uuid
        transfers:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BatchItemRequest"
    BatchItemRequest:
      type: object
      required: ["to_wallet_id", "asset", "amount"]
      properties:
        to_wallet_id:
          type: string
          format: uuid
        asset:
          type: string
        amount:
          type: string
          description: "Positive decimal amount"
        reference:
          type: string
          description: "Optional reference for this transfer"
    BatchResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: ["pending", "processing", "partial", "completed", "failed"]
        total_count:
          type: integer
        success_count:
          type: integer
        failed_count:
          type: integer
        created_at:
          type: string
          format: date-time
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/BatchTransferItem"
    BatchTransferItem:
      type: object
      properties:
        id:
          type: string
        to_wallet_id:
          type: string
        asset:
          type: string
        amount:
          type: string
        reference:
          type: string
          nullable: true
        status:
          type: string
          enum: ["pending", "submitted", "confirmed", "failed"]
        tx_hash:
          type: string
          nullable: true

    # ── FX ────────────────────────────────────────────────────────────────
...